software-device, -sensor
