+++ src/CMakeLists.txt
@@ -89,6 +89,8 @@ target_sources(${LRS_TARGET}
         "${CMAKE_CURRENT_LIST_DIR}/hid-sensor.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/uvc-sensor.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/software-device.cpp"
+        "${CMAKE_CURRENT_LIST_DIR}/software-device-info.cpp"
+        "${CMAKE_CURRENT_LIST_DIR}/software-sensor.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/source.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/stream.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/sync.cpp"
@@ -140,6 +142,8 @@ target_sources(${LRS_TARGET}
         "${CMAKE_CURRENT_LIST_DIR}/hid-sensor.h"
         "${CMAKE_CURRENT_LIST_DIR}/uvc-sensor.h"
         "${CMAKE_CURRENT_LIST_DIR}/software-device.h"
+        "${CMAKE_CURRENT_LIST_DIR}/software-device-info.h"
+        "${CMAKE_CURRENT_LIST_DIR}/software-sensor.h"
         "${CMAKE_CURRENT_LIST_DIR}/source.h"
         "${CMAKE_CURRENT_LIST_DIR}/stream.h"
         "${CMAKE_CURRENT_LIST_DIR}/sync.h"
